Add include_pages_meta config.
